refactor: Make anonymous classes into actual test asset classes
Required to allow us to test against PHP 5.6.
